requirements/dev.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
Flask-Cors==3.0.8
gunicorn==20.0.4
redis==3.5.3
Flask-APScheduler==1.11.0
pymysql==0.9.3
requests==2.23.0
validator.py==1.3.0